Do not use 'Deny' lint classifier
We use clippy for linting the compilation of various crates, some of
them being created by 3rd party. With recent versions clippy fails to
build some of those crates due to lints being classified as 'Deny'. The
code of those crates has not changed and is known to be working,
indicating that a compilation error is too strong of a mechanism to
signal a problem here.
Hence, with this change we avoid using the 'Deny' classifier for lints
that fall into the correctness category in favor of 'Warn'.